Fix memory and segment table problem
Function 'init_mem(void)' is not called when starting the simulation. A
call to that is now added to 'os.c'.

Segment table entry is not initialized with any value on the creation on
create. A new function 'init_segment_table(...)' is added and called in
the load function to initialize the newly created segment table.

Also add 'init_page_table(...)' for later use.